df_game_r: 快验证脚本 game_func_report.py + 正向修复构建 + 取消子代理遗留缺口
- 新增 game_func_report.py：单TU编译(不链接)+读 docs/class_func_reports oracle+生成合并报告
  (ORIG asm/Ghidra C/我们 asm/diff/分类)；--no-compile <1.5s；错拼 mangled 用 --class-override/--method。
- 新增 game_verify_guide.md：固化 4 条核心要求(快脚本验证/只用 class_func_reports/不编造不存在符号/
  参数类型可改/padding→真实字段 getter/setter 权威)。
- 正向修复被取消子代理破坏：advancealtar::ActionDefine/TimeLine、item_lock::stItemLockRef/CItemLock、
  stIndependentDropInfo/BuyShopData 重复定义→权威统一；CerashopAddRestrict/ServerParameterScript
  签名按 ORIG 修正；cUserHistoryLog::EnterDungeon 补 2 参；advancealtar::TimeLine 定义移入构建内；
  stDeathTower_t/live_server_info_t ctor 补齐。
- FixUndefinedSymbols：18 个真实 ORIG 符号补定义 + 3 个发明符号按 ORIG 真名修正调用
  (CUserGlobalInfoHandleIns→Instance()、STExpertJobScri::operator->→STExpertJobScript::GetLevel、DB_* const)。
- reset_uniqueid_flag asm 17→19 长度；SendCreatureItemList 迁 CCreatureMgr.cpp。
- compare_game_full：共同 19274 / identical 12636 / diff 5503；构建 0 错误 ELF。

+++ dnf_decompile/source/DNFServer/GameServer/Game/ServerParameterScript.cpp
@@ -380,14 +380,14 @@ int ServerParameterScript::GetAdmissionIncrease(int idx)
     return m_vec7a0[idx].second;
 }
 
-int ServerParameterScript::GetPvPPenaltyRevision(int grade)
+float ServerParameterScript::GetPvPPenaltyRevision(int grade)
 {
     std::map<int, int>::const_iterator it = m_mapb0.find(grade);
     if (it == m_mapb0.end())
     {
-        return 0;
+        return 0.0f;
     }
-    return it->second;
+    return (float)it->second;
 }
 
 int ServerParameterScript::getRecoverStaminaPrice(int idx, int price)
